backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-storages==1.14.2
boto3==1.29.7
Pillow==10.1.0
requests==2.31.0
python-decouple==3.8
drf-spectacular==0.26.5
django-filter==23.5
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
markdown==3.5.1
django-markdownify==0.9.2

